fix: address Sourcery review 4458966026 items
- setup_state CRUD consolidated in queries module (db.py re-exports)
- _TRANSITIONABLE_FIELDS constant removed (no longer needed)
- transition_state simplified to explicit SET clauses
- db.py comment updated (_UPDATABLE_FIELDS -> hardcoded SET clauses)
- JWKS test renamed to assert caching semantics explicitly
- SAFETY comments updated to reference hardcoded SET clauses
