+++ src/auth/db.py
@@ -79,82 +79,12 @@ def get_connection(db_path: str | None = None) -> Generator[sqlite3.Connection, 
 
 
 # ---------------------------------------------------------------------------
-# setup_state CRUD (for OIDC elicitation state machine)
+# setup_state CRUD — re-exported from queries module (source of truth)
 # ---------------------------------------------------------------------------
 
-def create_setup_state(
-    oidc_key: str,
-    state: str,
-    phone_number: str | None = None,
-    metadata: dict | None = None,
-    db_path: str | None = None,
-) -> None:
-    """Create a new setup_state row."""
-    import json
-    from datetime import datetime, timezone
-    now = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-    with get_connection(db_path) as conn:
-        conn.execute(
-            "INSERT INTO setup_state (oidc_key, state, phone_number, metadata, retry_count, created_at, updated_at) "
-            "VALUES (?, ?, ?, ?, 0, ?, ?)",
-            (oidc_key, state, phone_number, json.dumps(metadata) if metadata else None, now, now),
-        )
-
-
-def get_setup_state(oidc_key: str, db_path: str | None = None) -> dict | None:
-    """Fetch setup_state row as dict, or None if not found."""
-    with get_connection(db_path) as conn:
-        row = conn.execute(
-            "SELECT oidc_key, state, phone_number, tg_code_hash, retry_count, metadata, created_at, updated_at "
-            "FROM setup_state WHERE oidc_key = ?",
-            (oidc_key,),
-        ).fetchone()
-        return dict(row) if row else None
-
-
-def update_setup_state(
-    oidc_key: str,
-    state: str,
-    phone_number: str | None = None,
-    tg_code_hash: str | None = None,
-    metadata: dict | None = None,
-    retry_count: int | None = None,
-    db_path: str | None = None,
-) -> None:
-    """Update state and related fields. Refreshes updated_at."""
-    import json
-    from datetime import datetime, timezone
-    now = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-    sets = ["state = ?", "updated_at = ?"]
-    params: list = [state, now]
-    if phone_number is not None:
-        sets.append("phone_number = ?")
-        params.append(phone_number)
-    if tg_code_hash is not None:
-        sets.append("tg_code_hash = ?")
-        params.append(tg_code_hash)
-    if metadata is not None:
-        sets.append("metadata = ?")
-        params.append(json.dumps(metadata))
-    if retry_count is not None:
-        sets.append("retry_count = ?")
-        params.append(retry_count)
-    params.append(oidc_key)
-    with get_connection(db_path) as conn:
-        # SAFETY: Column names come from _UPDATABLE_FIELDS (hardcoded set), never user input.
-        # Values are bound parameters. Sourcery false positive — whitelist prevents injection.
-        conn.execute(  # noqa: S608
-            f"UPDATE setup_state SET {', '.join(sets)} WHERE oidc_key = ?",
-            params,
-        )
-
-
-def expire_old_states(cutoff_iso: str, db_path: str | None = None) -> int:
-    """Mark sessions with updated_at before cutoff_iso as EXPIRED. Returns count updated."""
-    with get_connection(db_path) as conn:
-        cur = conn.execute(
-            "UPDATE setup_state SET state = 'EXPIRED' "
-            "WHERE updated_at < ? AND state NOT IN ('COMPLETED', 'FAILED', 'EXPIRED')",
-            (cutoff_iso,),
-        )
-        return cur.rowcount
+from src.auth.queries.setup_state import (
+    create_setup_state,
+    get_setup_state,
+    update_setup_state,
+    expire_old_states,
+)
